requirements.txt
boto==2.47.0
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-storages==1.5.2
idna==2.5
mysql-connector==2.1.4
mysqlclient==1.3.4
numpy==1.13.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
selenium==3.4.3
six==1.10.0
urllib3==1.21.1
openpyxl==2.4.8
xlrd==1.1.0
pip==9.0.1
pandas==0.20.2
Pillow==4.2.1
scipy==0.19.1
SQLAlchemy==1.1.14
XlsxWriter==0.9.8